Implemented getAgeDifferenceInDays All test cases Green!

+++ src/main/java/uk/co/blc_services/gumtree/AddressRepository.java
@@ -57,7 +57,7 @@ public interface AddressRepository {
 	 * @param person b
 	 * @return days between their dob
 	 */
-	public int getAgeDifferenceInDays(Person a, Person b);
+	public long getAgeDifferenceInDays(Person a, Person b);
